+++ app/models/Context.py
@@ -1,8 +1,7 @@
-from app.utils.helper_functions import *
+from app.utils import *
 from datetime import datetime, timedelta # for date parsing
-from app.utils.constants import *
 from typing import *
-from app.models.Dimension import Dimension
+from app.models import Dimension
 
 class Context:
     """ define xbrl context object """
