+++ app/models/NetPosition.py
@@ -1,7 +1,4 @@
-from app.models.Table import Table
-from app.models.Cell import Cell
-from app.models.Context import Context
-from app.models.Dimension import Dimension
+from app.models import Table, Cell, Context, Dimension
 from app.utils.helper_functions import *
 from typing import *
