+++ app/models/Table.py
@@ -1,11 +1,8 @@
-from app.models.Context import Context
-from app.models.Cell import Cell
-from app.utils.helper_functions import *
+from app.models import Context, Cell
+from app.utils import *
 import pandas as pd
-from app.utils.constants import *
 from typing import * # to specify funtion inputs and outputs
 from datetime import datetime # for date parsing
-
 
 class Table:
     """ Object to represent an Excel sheet tab (one table) """
