+++ app/static/css/style.css
@@ -56,10 +56,6 @@ header {
     display:none;
 }
 
-#spinner {
-    display: none;
-}
-
 .loader {
     border: 16px solid #f3f3f3;
     border-radius: 50%;
