+++ main.py
@@ -1,7 +1,7 @@
 """
 Initialize the Flask application and run the server.
 
-K. Wheelan 2024
+Last updated: K. Wheelan, April 2024
 """
 
 from app import app
